% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_deseq.R
\name{Batch_Deseq_difference}
\alias{Batch_Deseq_difference}
\title{Batch DESeq Difference.}
\usage{
Batch_Deseq_difference(exprSet, group, num, save_dir = "Alldiffenece")
}
\arguments{
\item{exprSet}{exp data}

\item{group}{group info of samples}

\item{num}{Number of repetitions per group}

\item{save_dir}{A path to save the results}
}
\value{
Difference expression files corresponding to different groupings
}
\description{
Useful to compute a batch of DEG matrix
}
\examples{
\dontrun{
Batch_Deseq_difference(exprSet=exp_data,group,3,save_dir = tempdir())
}
}
\seealso{
Other Batch DESeq: 
\code{\link{exp_data}},
\code{\link{group}}
}
\concept{Batch DESeq}
