% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_violin_plot.R
\name{box_violin_plot}
\alias{box_violin_plot}
\title{box_violin_plot.}
\usage{
box_violin_plot(name)
}
\arguments{
\item{name}{a title text for your plot}
}
\value{
a box and violin plot
}
\description{
绘制小提琴+箱型图.
}
\examples{
name='Folate biosynthesis'
png(filename = paste0(tempdir(),'/Folate biosynthesis.png'),width = 1080,height = 1440,res = 300)
box_violin_plot(name)
dev.off()
}
\seealso{
Other box_violin_plot: 
\code{\link{Pathway}},
\code{\link{group2}}
}
\concept{box_violin_plot}
