% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_to_rownames.R
\name{col_to_rownames}
\alias{col_to_rownames}
\title{col_to_rownames}
\usage{
col_to_rownames(data, rownames, filter = TRUE)
}
\arguments{
\item{data}{expression matrix with a column of gene name}

\item{rownames}{colname of gene name}

\item{filter}{if TRUE, duplicate genes will removed by order of expression value, when FALSE will add suffixes to retain duplicate names}
}
\value{
a rownamed expression matrix
}
\description{
useful when expression matrix need to be rownamed by genes name,
duplicate gene names are automatically removed.
}
\examples{
col_to_rownames(dup_data,"genes")
}
