% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dup_data}
\alias{dup_data}
\title{Example data of col_to_rownames}
\format{
A data frame with 10 rows and 5 variables:
\describe{
\item{genes}{repeated gene names}
\item{sample1}{samples expression value}
\item{sample2}{samples expression value}
\item{sample3}{samples expression value}
\item{sample4}{samples expression value}
...
}
}
\usage{
dup_data
}
\description{
A Expression matrix with repeated gene names and expression value.
}
\keyword{datasets}
