% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exp_data}
\alias{exp_data}
\title{Example data of Batch DESeq.}
\format{
A gene rownamed data frame with 1000 rows and 12 variables:
\describe{
\item{A_1_1_1}{samples}
\item{A_2_2_2}{samples}
\item{A_3_3_3}{samples}
\item{B_4_4_4}{samples}
\item{B_5_5_5}{samples}
\item{B_6_6_6}{samples}
\item{C_7_7_7}{samples}
\item{C_8_8_8}{samples}
\item{C_9_9_9}{samples}
\item{D_10_10_10}{samples}
\item{D_11_11_11}{samples}
\item{D_12_12_12}{samples}
...
}
}
\usage{
exp_data
}
\description{
A fake dataset containing 12 samples and 1000 Genes of Batch DESeq.
}
\seealso{
Other Batch DESeq: 
\code{\link{Batch_Deseq_difference}()},
\code{\link{group}}
}
\concept{Batch DESeq}
\keyword{datasets}
