% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getgene.R
\name{getgene}
\alias{getgene}
\title{Get Gene}
\usage{
getgene(data, col)
}
\arguments{
\item{data}{a data frame contains column of genes in list form}

\item{col}{column of genes list}
}
\value{
a character list contains all genes
}
\description{
extract the column in the list format of the data frame
}
\examples{
getgene(data = sea_data, col = "leadingEdge")
}
