% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group}
\alias{group}
\title{Example data of Batch DESeq.}
\format{
A character vector of 12 variables:
\describe{
\item{A}{group}
\item{B}{group}
\item{C}{group}
...
}
}
\usage{
group
}
\description{
A fake group info of samples of Batch DESeq.
}
\seealso{
Other Batch DESeq: 
\code{\link{Batch_Deseq_difference}()},
\code{\link{exp_data}}
}
\concept{Batch DESeq}
\keyword{datasets}
