% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{group2}
\alias{group2}
\title{Example data of box_violin_plot.}
\format{
A character vector of 21 variables:
\describe{
\item{Control}{group}
\item{AB}{group}
\item{AA}{group}
...
}
}
\usage{
group2
}
\description{
A fake group info of samples of box_violin_plot.
}
\seealso{
Other box_violin_plot: 
\code{\link{Pathway}},
\code{\link{box_violin_plot}()}
}
\concept{box_violin_plot}
\keyword{datasets}
