% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note_data.R
\name{note_data}
\alias{note_data}
\alias{mysave}
\alias{myload}
\title{Note Data.}
\usage{
mysave(..., remark = "NULL")

myload(...)
}
\arguments{
\item{...}{the names of the objects to be saved (as symbols or
    character strings).}

\item{remark}{add a message to it}
}
\value{
a Rdata with note message
}
\description{
add note messages to Rdata produced by \code{save()}.
}
\examples{
a <- c("a","b")
file = paste0(tempdir(),"test.Rdata")
mysave(a,remark = "This is test message",file = file)
myload(file)
}
\seealso{
See \code{\link[base]{save}}
function from package (base)

See \code{\link[base]{load}}
function from package (base)
}
