% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pathway}
\alias{Pathway}
\title{Example data of box_violin_plot.}
\format{
A gene rownamed data frame with 69 rows and 21 variables:
\describe{
\item{Control1}{number}
\item{Control2}{number}
\item{Control3}{number}
...
}
}
\usage{
Pathway
}
\description{
A fake dataset containing metadata.
}
\seealso{
Other box_violin_plot: 
\code{\link{box_violin_plot}()},
\code{\link{group2}}
}
\concept{box_violin_plot}
\keyword{datasets}
