% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sea_data}
\alias{sea_data}
\title{Example data of getgene.}
\format{
A data frame with 91 rows and 8 variables:
\describe{
\item{ES}{number}
\item{NES}{number}
\item{leadingEdge}{list}
\item{nMoreExtreme}{character}
\item{padj}{number}
\item{pathway}{character}
\item{pval}{number}
\item{size}{integer}
...
}
}
\usage{
sea_data
}
\description{
A fake dataset.
}
\keyword{datasets}
